+++ include/task.hpp
@@ -21,21 +21,21 @@ class Task
 {
 
 public:
-    Task(int task_id, CommandFunction func, std::vector<uint8_t>& data, Payload* payload, int priority=0, std::string name="");
+    Task(uint8_t task_id, CommandFunction func, std::vector<uint8_t>& data, Payload* payload, uint8_t priority=0, std::string name="");
     // ~Task() = default; // TODO
 
     void Execute();
     int GetAttempts() const;
-    int GetPriority() const;
-    int GetID() const;
+    uint8_t GetPriority() const;
+    uint8_t GetID() const;
     size_t GetDataSize() const;
     std::chrono::system_clock::time_point GetCreationTime() const;
 
 private:
 
 
-    int task_id; // Unique identifier for the task.
-    int priority; // Priority of the task.
+    uint8_t task_id; // Unique identifier for the task.
+    uint8_t priority; // Priority of the task.
     std::vector<uint8_t> data; // Data to be passed to the task's function.
     Payload* payload; // Payload object to be passed to the task's function.
     CommandFunction func; // task's function to be executed.
